+++ src/main-en/java/fri/shapesge/ImageData.java
@@ -1,0 +1,44 @@
+package fri.shapesge;
+
+import fri.shapesge.engine.Game;
+
+import java.awt.image.BufferedImage;
+
+/**
+ * The ImageData class represents a bitmap image that can be drawn using the {@link Image} class.
+ *
+ * @author Ján Janech
+ * @version 1.0
+ */
+public class ImageData {
+
+    private final BufferedImage image;
+
+    /**
+     * Reads an image from a given path.
+     * @param imagePath path to the image
+     */
+    public ImageData(String imagePath) {
+        this.image = Game.getGame().getParser().parseImage(imagePath);
+    }
+
+    /**
+     * Returns width of the image.
+     * @return width of the image
+     */
+    public int getWidth() {
+        return this.image.getWidth();
+    }
+
+    /**
+     * Returns height of the image.
+     * @return height of the image
+     */
+    public int getHeight() {
+        return this.image.getHeight();
+    }
+
+    BufferedImage getImage() {
+        return this.image;
+    }
+}